Refactor: Clean up pose estimator. Add more cameras to better reflect the real robot's condition.
