+++ src/main/java/frc/robot/RobotContainer.java
@@ -9,7 +9,7 @@ import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 import frc.lib.generic.GenericSubsystem;
 import frc.lib.util.Controller;
 import frc.lib.util.flippable.Flippable;
-import frc.robot.poseestimation.poseestimator.PoseEstimator5990;
+import frc.robot.poseestimation.poseestimator.PoseEstimator;
 import frc.robot.subsystems.leds.Leds;
 import frc.robot.subsystems.swerve.Swerve;
 import frc.robot.subsystems.swerve.SwerveCommands;
@@ -20,14 +20,14 @@ import java.util.function.DoubleSupplier;
 
 import static frc.lib.util.Controller.Axis.LEFT_X;
 import static frc.lib.util.Controller.Axis.LEFT_Y;
-import static frc.robot.poseestimation.poseestimator.PoseEstimatorConstants.FRONT_CAMERA;
-import static frc.robot.utilities.FieldLocations.BLUE_BOTTOM_FEEDER;
-import static frc.robot.utilities.FieldLocations.BLUE_TOP_FEEDER;
-import static frc.robot.utilities.PathPlannerConstants.PATHPLANNER_CONSTRAINTS;
+import static frc.robot.poseestimation.poseestimator.PoseEstimatorConstants.*;
 
 public class RobotContainer {
-    public static final PoseEstimator5990 POSE_ESTIMATOR = new PoseEstimator5990(
-            FRONT_CAMERA
+    public static final PoseEstimator POSE_ESTIMATOR = new PoseEstimator(
+            FRONT_LEFT_CAMERA,
+            FRONT_RIGHT_CAMERA,
+            REAR_LEFT_CAMERA,
+            REAR_RIGHT_CAMERA
     );
 
     public static final Swerve SWERVE = new Swerve();
@@ -55,13 +55,6 @@ public class RobotContainer {
         DriverStation.silenceJoystickConnectionWarning(true);
 
         setupLEDs();
-
-        driveController.getButton(Controller.Inputs.A).whileTrue(AutoBuilder.pathfindToPose(
-                BLUE_TOP_FEEDER.toPose2d(), PATHPLANNER_CONSTRAINTS));
-
-        driveController.getButton(Controller.Inputs.B).whileTrue(AutoBuilder.pathfindToPose(
-                BLUE_BOTTOM_FEEDER.toPose2d(), PATHPLANNER_CONSTRAINTS));
-
 
         configureButtons(ButtonLayout.TELEOP);
     }
