+++ src/main/java/frc/robot/subsystems/swerve/Swerve.java
@@ -47,7 +47,7 @@ public class Swerve extends GenericSubsystem {
             gyroRotations[i] = Rotation2d.fromDegrees(odometryUpdatesYawDegrees[i]);
         }
 
-        POSE_ESTIMATOR.updatePoseEstimatorStates(
+        POSE_ESTIMATOR.updateFromOdometry(
                         swerveWheelPositions,
                         gyroRotations,
                         OdometryThread.getInstance().getLatestTimestamps()
