+++ src/test/java/frc/robot/ButtonTest.java
@@ -1,5 +1,8 @@
 package frc.robot;
 
+import edu.wpi.first.apriltag.AprilTag;
+import edu.wpi.first.apriltag.AprilTagFieldLayout;
+import edu.wpi.first.apriltag.AprilTagFields;
 import edu.wpi.first.math.controller.HolonomicDriveController;
 import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.math.controller.ProfiledPIDController;
@@ -25,13 +28,12 @@ class ButtonTest {
 //
         Map<Integer, Pose3d> TAG_ID_TO_POSE = new HashMap<>();
 
-//        for (AprilTag aprilTag : AprilTagFields.k2024Crescendo.loadAprilTagLayoutField().getTags())
-//            TAG_ID_TO_POSE.put(aprilTag.ID, aprilTag.pose);
-
+        for (AprilTag aprilTag : AprilTagFieldLayout.loadField(AprilTagFields.k2025Reefscape).getTags())
+            TAG_ID_TO_POSE.put(aprilTag.ID, aprilTag.pose);
 
         for (Pose3d tagPosition : TAG_ID_TO_POSE.values()) {
-            System.out.println("Y:" + TAG_ID_TO_POSE.get(7).getY()); //Y:5.547868
-            System.out.println("X:" + TAG_ID_TO_POSE.get(7).getX()); //X:-0.0381
+            System.out.println("Y:" + TAG_ID_TO_POSE.get(tagPosition).getY()); //Y:5.547868
+            System.out.println("X:" + TAG_ID_TO_POSE.get(tagPosition).getX()); //X:-0.0381
         }
 //
     }
